refactor(core): move _is_numeric_dtype into _array_backend (arob5 review)
Addresses arob5's review on PR #365. _is_numeric_leaf did a per-call
`from .event_template import _is_numeric_dtype` on the duck path -- a hot path
(NumericRecord construction gate, workflow-result coercion, template inference)
that bare numpy / jax arrays always take. The lazy import existed only to dodge
the _array_backend -> event_template cycle.

Move _is_numeric_dtype and _NUMERIC_KINDS down into _array_backend, co-located
with the leaf resolvers that wrap the predicate. This removes the per-call
import (three lazy imports in _array_backend, including both _register_* hooks)
and the upward dependency edge -- clean DAG: _array_backend <- event_template
<- record <- _numeric_record.

Its historical importers (_record_array, _broadcast_distributions,
record.design, and the _fingerprint lazy import) now import _is_numeric_dtype
directly from _array_backend rather than via an event_template re-export, so
event_template no longer carries it at all. test_numeric_dtype_predicate_shared
pins the new home. Full tests/core plus distributions / converters / modeling /
inference / diagnostics / validation green.
